fix(vue3): Apply visible prop last for dialog components
Other props such as animation and maskColor need to be set before the
dialog is shown, so this ensures they are set properly first.
