+++ vue3-onsenui/src/components/VOnsActionSheet.vue
@@ -1,5 +1,5 @@
 <template>
-  <ons-action-sheet :modifier="normalizedModifier">
+  <ons-action-sheet :modifier="normalizedModifier" v-bind="orderedProps">
     <slot></slot>
   </ons-action-sheet>
 </template>
@@ -7,12 +7,12 @@
 <script>
   /* This file was generated automatically by 'generate-components' task in bindings/vue/gulpfile.babel.js */
   import 'onsenui/esm/elements/ons-action-sheet.js';
-  import { deriveEvents, hasOptions, dialogCancel, deriveDBB, portal, modifier } from '../mixins/index.js';
+  import { deriveEvents, hasOptions, dialog, deriveDBB, portal, modifier } from '../mixins/index.js';
 
   const name = 'v-ons-action-sheet';
 
   export default {
     name,
-    mixins: [deriveEvents(name), hasOptions, dialogCancel, deriveDBB, portal, modifier]
+    mixins: [deriveEvents(name), hasOptions, dialog, deriveDBB, portal, modifier]
   };
 </script>
