+++ vue3-onsenui/src/components/VOnsDialog.vue
@@ -1,5 +1,5 @@
 <template>
-  <ons-dialog :modifier="normalizedModifier">
+  <ons-dialog :modifier="normalizedModifier" v-bind="orderedProps">
     <slot></slot>
   </ons-dialog>
 </template>
@@ -7,12 +7,12 @@
 <script>
   /* This file was generated automatically by 'generate-components' task in bindings/vue/gulpfile.babel.js */
   import 'onsenui/esm/elements/ons-dialog.js';
-  import { deriveEvents, hasOptions, dialogCancel, deriveDBB, portal, modifier } from '../mixins/index.js';
+  import { deriveEvents, hasOptions, dialog, deriveDBB, portal, modifier } from '../mixins/index.js';
 
   const name = 'v-ons-dialog';
 
   export default {
     name,
-    mixins: [deriveEvents(name), hasOptions, dialogCancel, deriveDBB, portal, modifier]
+    mixins: [deriveEvents(name), hasOptions, dialog, deriveDBB, portal, modifier]
   };
 </script>
