+++ vue3-onsenui/src/mixins/index.js
@@ -1,3 +1,4 @@
 export { deriveDBB, deriveEvents, unrecognizedListeners } from './derive';
-export { hidable, hasOptions, selfProvider, dialogCancel, portal, modifier } from './common';
+export { hidable, hasOptions, selfProvider, portal, modifier } from './common';
 export { modelInput, modelInputNumber, modelCheckbox, modelRadio } from './model';
+export { dialog } from './dialog';
